+++ packages/package_info/lib/package_info.dart
@@ -32,6 +32,9 @@ class PackageInfo {
     if (_fromPlatform == null) {
       final Completer<PackageInfo> completer = Completer<PackageInfo>();
 
+      // TODO(amirh): remove this on when the invokeMethod update makes it to stable Flutter.
+      // https://github.com/flutter/flutter/issues/26431
+      // ignore: strong_mode_implicit_dynamic_method
       _kChannel.invokeMethod('getAll').then((dynamic result) {
         final Map<dynamic, dynamic> map = result;