+++ packages/share/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.6.0+1
+
+* Log a more detailed warning at build time about the previous AndroidX
+  migration.
+
+## 0.6.0
+
+* **Breaking change**. Migrate from the deprecated original Android Support
+  Library to AndroidX. This shouldn't result in any functional changes, but it
+  requires any Android apps using this plugin to [also
+  migrate](https://developer.android.com/jetpack/androidx/migrate) if they're
+  using the original support library.
+
 ## 0.5.3
 
 * Added missing test package dependency.